Move IReference ABI to WindowsRuntime_ABI.
